Remove the already initialized constant error
emopop:chef-repo(master)% spiceweasel infrastructure.yml
/Users/jdewey/.rvm/gems/ruby-1.9.3-p194@chef-repo/gems/spiceweasel-1.1.2/bin/spiceweasel:33: warning: already initialized constant DEBUG

Tracked this down to bundler's no_exec_wrapper.
It defines DEBUG on the global namespace.
  https://github.com/mpapis/rubygems-bundler/blob/master/lib/rubygems-bundler/noexec.rb

Maybe bundler is in the wrong for defining this on the global namespace.
However, we are doing the same thing in spiceweasel.  Decided to namespace
all the constants under ::Spiceweasel.